Add a patch to allow searching in user auths in the admin
Searching in the user social auth section is required as people
sometimes report issues with their auth provider not being associated
with their ambassador profile and manual editing via the admin is
possible only if the corresponding social auth is found.

The patch is necessary because our versions of both python-social-auth
and Django are too old and the auto-discovery of search fields is not
possible. Our version of python-social-auth also has a bug in its search
implementation so I just manually define the search fields in a
monkey-patch style.

This should be removed if we upgrade Django and python-social-auth to
newer versions.