Exclude Antarctica and rename Taiwan
Antarctica is not a popular destination for coding events. But who
knows--we might have one someday and we can put it back there.

Taiwan was renamed from "Taiwan (A Provice of China)" to just "Taiwan".
It seems that the official name as defined by ISO 3166-1 is the longer
one, but some Taiwanese find it offensive and it's better to leave it to
just Taiwan.

See here for more info:
https://github.com/SmileyChris/django-countries/issues/29