Temporary hide the maps until they're optimized
Right now the site is super slow and almost never loads because of the
many events for this year's edition. The map tries to load over 6000
events via JSON and the app returns a 2+ MB JSON file. That's slow to
generate, slow to cache, slow to transfer and slow to process. The
smallish VPS we use can't handle it well.

That's why I'm hiding the map so that at least the other parts of the
site can work. When someone has some time to address the problem, this
commit should be reverted.