Show only countries from the current year by default
We were fetching all events from the DB before, to show a list of
countries. That's pretty slow. We now do this only if the "Show past
events" option is selected by the user.

The method still needs to be optimized a lot, since it loads all events,
then traverses them in the code just to fetch their countries.