Optimize the list_active_countries() helper
Previously this function was fetching all events with a few columns (not
all, but ID, start_date and country at least). If we have more than a
few hundred events (as we have now--a couple thousand) this becomes way
too slow and memory hungry. It's best if we let the DB do the hard work
and return us a list of only the country codes of all current events
(which would be just a few dozen entries tops). We then build the list
of countries using Django's countries list plugin.