Allow picking an event location by clicking on the map
Previously it was possible to drag the location's marker and change the
location in this way, but the marker has to be on the map already. A
marker appears only after a successful address geocode. If the geocoding
fails, no marker would be present and there would be no way to set a
location.

People should now be able to just click on the map to set the location.

Also, when clicking “Enter” in the location field, the page will no
longer be reloaded.